refactor: Code review notes applied
