+++ packages/cli/src/commands/__tests__/generate/index.ts
@@ -11,8 +11,8 @@ import symlinkDir from "symlink-dir";
 import tempDir from "temp-dir";
 import tmp from "tmp-promise";
 import { promisify } from "util";
-import { setLogger } from "../../cli/command";
-import { GenerateArguments, generateCmd } from "../generate";
+import { setLogger } from "../../../cli/command";
+import { GenerateArguments, generateCmd } from "../../generate";
 
 jest.mock("@kosko/generate");
 jest.mock("@kosko/env");
@@ -243,8 +243,8 @@ describe("with components in config", () => {
       args.set = [{ key: "a", value: "b" }];
     });
 
-    test("should add variables layer to the env", () => {
-      expect(env.addVariablesLayer).toHaveBeenCalledWith(expect.any(Function));
+    test("should add reducer to the env", () => {
+      expect(env.setReducers).toHaveBeenCalledWith(expect.any(Function));
     });
   });
